- switching to gcm and the cipher package for encryption
- switching to hex encoding (base16) vs base64

=== OLD ===
BenchmarkEncryptDataBlock-8       500000        2585 ns/op
BenchmarkEncodeText-8             500000        3436 ns/op
BenchmarkDecodeText-8             500000        2911 ns/op
BenchmarkContainsSubString-8    200000000          8.85 ns/op

=== NEW ===
BenchmarkEncryptDataBlock-8    	  500000	      2202 ns/op
BenchmarkEncodeText-8          	  500000	      3343 ns/op
BenchmarkDecodeText-8          	 1000000	      2130 ns/op
BenchmarkContainsSubString-8   	200000000	         8.25 ns/op